pyproject.toml
[project]
name = "distlisp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "uvicorn>=0.43.0",
]
